[REM] l10n_br_fiscal: remove document.move.mixin
- copy the related fields directly into document.py
- copy _onchange_document_type_id and _onchange_document_serie_id to l10n_br_fiscal/models/document_mixin_methods.py
