+++ l10n_br_fiscal/models/__init__.py
@@ -6,7 +6,6 @@ from . import data_ncm_nbs_abstract
 from . import document_mixin_methods
 from . import document_mixin_fields
 from . import document_mixin
-from . import document_move_mixin
 from . import document_line_mixin_methods
 from . import document_line_mixin
 from . import invalidate_number
